requirements.txt
attrs==18.2.0
click==6.7
colorama==0.3.9
pyperclip==1.6.4
PyQt5==5.10.1
PySide2==5.11.1
Qt.py==1.1.0
sip==4.19.8
ujson==1.35
